+++ apps/web/src/views/Farms/components/FarmCard/CardActionsContainer.tsx
@@ -1,13 +1,13 @@
-import { ChainId } from '@pancakeswap/chains'
 import { FarmWithStakedValue } from '@pancakeswap/farms'
 import { useTranslation } from '@pancakeswap/localization'
 import { AtomBox, Button, Flex, RowBetween, Skeleton, Text } from '@pancakeswap/uikit'
 import ConnectWalletButton from 'components/ConnectWalletButton'
-import { useActiveChainId } from 'hooks/useActiveChainId'
-import NextLink from 'next/link'
+
+import { useRouter } from 'next/router'
 import { useMemo } from 'react'
 import { styled, useTheme } from 'styled-components'
 import { StatusView } from 'views/Farms/components/YieldBooster/components/bCakeV3/StatusView'
+import { StatusViewButtons } from 'views/Farms/components/YieldBooster/components/bCakeV3/StatusViewButtons'
 import { useBCakeBoostLimitAndLockInfo } from 'views/Farms/components/YieldBooster/hooks/bCakeV3/useBCakeV3Info'
 import { useBoostStatusPM } from 'views/Farms/components/YieldBooster/hooks/bCakeV3/useBoostStatus'
 import { useWrapperBooster } from 'views/PositionManagers/hooks'
@@ -74,7 +74,8 @@ const CardActions: React.FC<React.PropsWithChildren<FarmCardActionsProps>> = ({
   )
   const { onUpdate } = useUpdateBCakeFarms(bCakeWrapperAddress ?? '0x', pid)
   const { locked } = useBCakeBoostLimitAndLockInfo()
-  const { chainId } = useActiveChainId()
+  const router = useRouter()
+  const isHistory = useMemo(() => router.pathname.includes('history'), [router])
 
   return (
     <AtomBox mt="16px">
@@ -131,7 +132,7 @@ const CardActions: React.FC<React.PropsWithChildren<FarmCardActionsProps>> = ({
             </RowBetween>
           </>
         )}
-        {isBooster && chainId === ChainId.BSC && (
+        {isBooster && !isHistory && (
           <>
             <AtomBox
               width={{
@@ -150,16 +151,14 @@ const CardActions: React.FC<React.PropsWithChildren<FarmCardActionsProps>> = ({
                   shouldUpdate={shouldUpdate && farm?.bCakeUserData?.stakedBalance?.gt(0)}
                   expectMultiplier={veCakeUserMultiplierBeforeBoosted}
                 />
-                {!locked && (
-                  <NextLink href="/cake-staking" passHref>
-                    <Button width="100%" style={{ whiteSpace: 'nowrap' }}>
-                      {t('Go to Lock')}
-                    </Button>
-                  </NextLink>
-                )}
-                {shouldUpdate && farm?.bCakeUserData?.stakedBalance?.gt(0) && (
-                  <Button onClick={onUpdate}>{t('Update')}</Button>
-                )}
+                <StatusViewButtons
+                  locked={locked}
+                  updateButton={
+                    shouldUpdate && farm?.bCakeUserData?.stakedBalance?.gt(0) ? (
+                      <Button onClick={onUpdate}>{t('Update')}</Button>
+                    ) : null
+                  }
+                />
               </Flex>
             </RowBetween>
           </>