+++ apps/web/src/views/PositionManagers/components/LiquidityManagement.tsx
@@ -9,13 +9,13 @@ import { useCurrencyBalances } from 'state/wallet/hooks'
 import { Address } from 'viem'
 
 import ConnectWalletButton from 'components/ConnectWalletButton'
-import NextLink from 'next/link'
 import { styled, useTheme } from 'styled-components'
 import { StatusView } from 'views/Farms/components/YieldBooster/components/bCakeV3/StatusView'
+import { StatusViewButtons } from 'views/Farms/components/YieldBooster/components/bCakeV3/StatusViewButtons'
 import { useBCakeBoostLimitAndLockInfo } from 'views/Farms/components/YieldBooster/hooks/bCakeV3/useBCakeV3Info'
 import { useBoostStatusPM } from 'views/Farms/components/YieldBooster/hooks/bCakeV3/useBoostStatus'
 import { useAccount } from 'wagmi'
-import { AprDataInfo, useWrapperBooster } from '../hooks'
+import { AprDataInfo, usePMV2SSMaxBoostMultiplier, useWrapperBooster } from '../hooks'
 import { useOnStake } from '../hooks/useOnStake'
 import { AddLiquidity } from './AddLiquidity'
 import { RemoveLiquidity } from './RemoveLiquidity'
@@ -152,6 +152,7 @@ export const LiquidityManagement = memo(function LiquidityManagement({
 
   const showRemoveLiquidityModal = useCallback(() => setRemoveLiquidityModalOpen(true), [])
   const hideRemoveLiquidityModal = useCallback(() => setRemoveLiquidityModalOpen(false), [])
+  const { maxBoostMultiplier } = usePMV2SSMaxBoostMultiplier()
 
   const relevantTokenBalances = useCurrencyBalances(
     account ?? undefined,
@@ -223,18 +224,16 @@ export const LiquidityManagement = memo(function LiquidityManagement({
                       status={status}
                       isFarmStaking
                       boostedMultiplier={boosterMultiplier}
-                      maxBoostMultiplier={3}
+                      maxBoostMultiplier={maxBoostMultiplier}
                       shouldUpdate={shouldUpdate}
                       expectMultiplier={veCakeUserMultiplierBeforeBoosted}
                     />
-                    {shouldUpdate && <Button onClick={() => onUpdate(refetch)}>{t('Update')}</Button>}
-                    {!locked && (
-                      <NextLink href="/cake-staking" passHref>
-                        <Button width="100%" style={{ whiteSpace: 'nowrap' }}>
-                          {t('Go to Lock')}
-                        </Button>
-                      </NextLink>
-                    )}
+                    <StatusViewButtons
+                      updateButton={
+                        shouldUpdate ? <Button onClick={() => onUpdate(refetch)}>{t('Update')}</Button> : null
+                      }
+                      locked={locked}
+                    />
                   </Flex>
                 </RowBetween>
               </>
@@ -265,14 +264,8 @@ export const LiquidityManagement = memo(function LiquidityManagement({
                 />
                 <RowBetween flexDirection="column" alignItems="flex-start" flex={1} width="100%">
                   <Flex width="100%" justifyContent="space-between" alignItems="center">
-                    <StatusView status={status} maxBoostMultiplier={3} />
-                    {!locked && (
-                      <NextLink href="/cake-staking" passHref>
-                        <Button width="100%" style={{ whiteSpace: 'nowrap' }}>
-                          {t('Go to Lock')}
-                        </Button>
-                      </NextLink>
-                    )}
+                    <StatusView status={status} maxBoostMultiplier={maxBoostMultiplier} />
+                    <StatusViewButtons updateButton={null} locked={locked} />
                   </Flex>
                 </RowBetween>
               </>