+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pancakeswap/tokens
 
+## 0.6.21
+
+### Patch Changes
+
+- Updated dependencies [9a16780]
+  - @pancakeswap/chains@0.4.6
+  - @pancakeswap/sdk@5.8.7
+  - @pancakeswap/token-lists@0.0.11
+
 ## 0.6.20
 
 ### Patch Changes