+++ packages/v3-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pancakeswap/v3-sdk
 
+## 3.8.8
+
+### Patch Changes
+
+- Updated dependencies [9a16780]
+  - @pancakeswap/chains@0.4.6
+  - @pancakeswap/sdk@5.8.7
+
 ## 3.8.7
 
 ### Patch Changes